tests/wpt/css-tests/css-flexbox-1_dev/html/flexbox_computedstyle_min-width-auto.htm

<!DOCTYPE html>
<html><head><title>flexbox | computed style | min-width: auto</title>
<link href="http://opera.com" rel="author" title="Opera Software">
<link href="http://www.w3.org/TR/css-flexbox-1/#min-size-auto" rel="help">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta content="dom" name="flags">
<style>
html {
	display: flex;
}
body {
	color: red;
	min-width: auto;
}
</style>
</head><body><div id="log"></div>
<script>
test(function() {
    var body = document.body;

    assert_equals(getComputedStyle(body).getPropertyValue("min-width"),
	          "auto");
});
</script>
</body></html>
